old_python_fastapi_backend/pyproject.toml
[tool.poetry]
name = "saferoute"
version = "0.1.0"
description = "Safety-scored routing service using UK Police crime data"
authors = ["SafeRoute Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.0"
geoalchemy2 = "^0.15.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.12.0"}
pydantic-settings = "^2.5.0"
email-validator = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
argon2-cffi = "^23.1.0"
httpx = "^0.27.0"
redis = "^5.0.0"
shapely = "^2.0.0"
pyproj = "^3.6.0"
python-multipart = "^0.0.9"
gunicorn = "^22.0.0"
slowapi = "^0.1.9"
python-json-logger = "^2.0.7"
python-dateutil = "^2.9.0"
celery = "^5.3.0"
h3 = "^4.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
black = "^24.8.0"
isort = "^5.13.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
testcontainers = "^4.8.0"
httpx = "^0.27.0"
faker = "^30.0.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

